chtml/html/testdata/webkit/tests5.dat

#data
<style> <!-- </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   " <!-- "
| "x"

#data
<style> <!-- </style> --> </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,34): unexpected-end-tag
#document
| <style>
|   " <!-- "
| " --> x"

#data
<style> <!--> </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   " <!--> "
| "x"

#data
<style> <!---> </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   " <!---> "
| "x"

#data
<iframe> <!---> </iframe>x
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <iframe>
|   " <!---> "
| "x"

#data
<iframe> <!--- </iframe>->x</iframe> --> </iframe>x
#errors
(1,8): expected-doctype-but-got-start-tag
(1,36): unexpected-end-tag
(1,50): unexpected-end-tag
#document
| <iframe>
|   " <!--- "
| "->x --> x"

#data
<script> <!-- </script> --> </script>x
#errors
(1,8): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <script>
|   " <!-- "
| " --> x"

#data
<title> <!-- </title> --> </title>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,34): unexpected-end-tag
#document
| <title>
|   " <!-- "
| " --> x"

#data
<textarea> <!--- </textarea>->x</textarea> --> </textarea>x
#errors
(1,10): expected-doctype-but-got-start-tag
(1,42): unexpected-end-tag
(1,58): unexpected-end-tag
#document
| <textarea>
|   " <!--- "
| "->x --> x"

#data
<style> <!</-- </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <style>
|   " <!</-- "
| "x"

#data
<p><xmp></xmp>
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <p>
| <xmp>

#data
<xmp> <!-- > --> </xmp>
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <xmp>
|   " <!-- > --> "

#data
<title>&amp;</title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <title>
|   "&"

#data
<title><!--&amp;--></title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <title>
|   "<!--&-->"

#data
<title><!--</title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <title>
|   "<!--"

#data
<noscript><!--</noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,39): unexpected-end-tag
#document
| <noscript>
|   <!-- </noscript> -->

#data
<noscript><!--</noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <noscript>
|   <!-- </noscript> -->
